Mesh updating with noise values
The noise textures are now retrieved from the GPU and the mesh is updated using these values. It needs to be updated to use a more efficient way of reading the colour values.

+++ Assets/Scripts/NoiseGenerator.cs
@@ -19,7 +19,7 @@ public class NoiseGenerator : Singleton<NoiseGenerator>
     {
         UpdateNoiseValues();
         float chunkSize = TerrainGenerationManager.Instance.ChunkSize;
-        test.SetFloats("_ChunkCoords", (float)chunkCoords.x, (float)chunkCoords.y);
+        test.SetFloats("_ChunkCoords", chunkCoords.x, chunkCoords.y);
         test.SetFloat("_ChunkSize", TerrainGenerationManager.Instance.VerticesAlongEdge);
         test.SetTexture(0, "_ChunkHeightMap", heightMapToWrite);
         test.Dispatch(0, heightMapToWrite.width / 8, heightMapToWrite.height / 8, 1);
